# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/v1/expenses/update": {
            "put": {
                "tags": [
                    "Expense",
                    "UPDATE"
                ],
                "operationId": "updateExpense",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateExpenseDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseExpenseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/expenses/save": {
            "post": {
                "tags": [
                    "Expense",
                    "SAVE"
                ],
                "operationId": "saveExpense",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateExpenseDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseExpenseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/expenses/range/{pageNumber}/{pageSize}": {
            "get": {
                "tags": [
                    "Expense",
                    "FIND"
                ],
                "operationId": "getExpensesInRange",
                "parameters": [
                    {
                        "name": "startDate",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "endDate",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pageNumber",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ResponseExpenseDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/expenses/delete/{id}": {
            "delete": {
                "tags": [
                    "DELETE",
                    "Expense"
                ],
                "operationId": "updateExpense_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UpdateExpenseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "categoryId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "description": {
                        "type": "string"
                    },
                    "profileId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "accountId": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "required": [
                    "id"
                ]
            },
            "Account": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "accountName": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "accountBalance": {
                        "type": "number"
                    },
                    "accountType": {
                        "type": "string",
                        "enum": [
                            "SAVINGS",
                            "CURRENT"
                        ]
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "expenses": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Expense"
                        },
                        "uniqueItems": true
                    },
                    "profile": {
                        "$ref": "#/components/schemas/Profile"
                    }
                }
            },
            "Category": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "categoryName": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "colorCode": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "expenses": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Expense"
                        },
                        "uniqueItems": true
                    },
                    "profile": {
                        "$ref": "#/components/schemas/Profile"
                    }
                }
            },
            "Expense": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "description": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "account": {
                        "$ref": "#/components/schemas/Account"
                    },
                    "profile": {
                        "$ref": "#/components/schemas/Profile"
                    }
                }
            },
            "GrantedAuthority": {
                "type": "object",
                "properties": {
                    "authority": {
                        "type": "string"
                    }
                }
            },
            "Profile": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "profileName": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "colorCode": {
                        "type": "string"
                    },
                    "profilePicture": {
                        "type": "string",
                        "format": "byte"
                    },
                    "profilePictureExtension": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "accounts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Account"
                        },
                        "uniqueItems": true
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        },
                        "uniqueItems": true
                    },
                    "expenses": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Expense"
                        },
                        "uniqueItems": true
                    }
                }
            },
            "ResponseExpenseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "description": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date"
                    },
                    "profileId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "accountId": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "required": [
                    "accountId",
                    "amount",
                    "category",
                    "createdAt",
                    "description",
                    "id",
                    "profileId",
                    "updatedAt"
                ]
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "ADMIN",
                            "USER"
                        ]
                    },
                    "expenses": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Expense"
                        },
                        "uniqueItems": true
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        },
                        "uniqueItems": true
                    },
                    "accounts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Account"
                        },
                        "uniqueItems": true
                    },
                    "profiles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Profile"
                        },
                        "uniqueItems": true
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "credentialsNonExpired": {
                        "type": "boolean"
                    },
                    "accountNonExpired": {
                        "type": "boolean"
                    },
                    "authorities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GrantedAuthority"
                        }
                    },
                    "accountNonLocked": {
                        "type": "boolean"
                    }
                }
            },
...
            "CreateExpenseDTO": {
                "type": "object",
                "properties": {
                    "categoryId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "description": {
                        "type": "string"
                    },
                    "profileId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "accountId": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "required": [
                    "accountId",
                    "categoryId",
                    "profileId"
                ]
            },
...
        }
    }
}